-- moved delineariztor -- moved IndiRef -- created static.cc for precompiled headers -- moved some code to utils -- fixed includes -- fixed observers -- fixed stl iterator usage (erase) -- observer debug simplified (perhaps too much? :))

+++ src/kernel/cpdf.h
@@ -227,7 +227,7 @@ public:
 
 	/** Type for pdf identificator.
 	 */
-	typedef unsigned long cpdf_id_t;
+	typedef unsigned long long cpdf_id_t;
 
 	/** Constant for pdf id of no pdf.
 	 * This is used for properties which comes from no pdf. Each CPdf instance
